+++ src/formatter/EpubFormatter.py
@@ -1,0 +1,5 @@
+import src.formatter.BasicFormatter as bf
+
+
+def run(input_markdown_path: str):
+    return bf.run(input_markdown_path)
